[5] split lofs of code in same file
