+++ lab5/src/thread.c
@@ -1,123 +1,22 @@
 #include "thread.h"
 #include "mini_uart.h"
 #include "timer.h"
+#include "sche.h"
+#include "list.h"
 #include "mm.h"
 #include "exception_c.h"
 #include "current.h"
 
-pid_t task_count = 0;
-
-list running_queue = LIST_HEAD_INIT(running_queue);
-list waiting_queue = LIST_HEAD_INIT(waiting_queue);
-list stopped_queue = LIST_HEAD_INIT(stopped_queue);
-
-////////////////////////////////////////////////////////////////////////////////////////////////
-//                                       schedule                                             //
-////////////////////////////////////////////////////////////////////////////////////////////////
-struct pt_regs
+void thread_kill_zombies()
 {
-    unsigned long regs[31];
-    unsigned long sp;
-    unsigned long pc;
-    unsigned long pstate;
-};
-
-int get_the_cur_count()
-{
-    int count = 0;
-    list *head = &running_queue;
-    while (head->next != &running_queue)
+    unsigned long flags = disable_irq();
+    while (!list_empty(&stopped_queue))
     {
-        count++;
-        head = head->next;
+        struct task *victim = list_first_entry(&stopped_queue, struct task, list);
+        unlink(&victim->list);
+        free_task(victim);
     }
-    return count;
-}
-
-void add_task(struct task *t)
-{
-    size_t flags = disable_irq();
-    insert_tail(&running_queue, &t->list);
     irq_restore(flags);
-}
-
-void kill_task(struct task *_task, int status)
-{
-    size_t flags = disable_irq();
-
-    _task->state = TASK_STOPPED;
-    _task->need_resched = 1;
-    unlink(&_task->list);
-
-    _task->exitcode = status;
-    insert_head(&stopped_queue, &_task->list);
-
-    irq_restore(flags);
-}
-
-void restart_task(struct task *_task)
-{
-    size_t flags = disable_irq();
-
-    if (_task->state != TASK_WAITING)
-    {
-        uart_send_string("---- task state inconsistent ----\n");
-        return;
-    }
-    _task->state = TASK_RUNNING;
-    unlink(&_task->list);
-    insert_tail(&running_queue, &_task->list);
-    irq_restore(flags);
-}
-
-void pause_task(struct task *_task)
-{
-    size_t flags = disable_irq();
-
-    if (_task->state != TASK_RUNNING)
-    {
-        uart_send_string("---- task state inconsistent ----\n");
-        return;
-    }
-    _task->state = TASK_WAITING;
-    _task->need_resched = 1;
-    unlink(&_task->list);
-    insert_head(&waiting_queue, &_task->list);
-
-    irq_restore(flags);
-}
-
-void sleep_task(size_t ms)
-{
-    size_t flags = disable_irq();
-    add_timer((timer_callback)restart_task, (size_t)current, ms);
-    pause_task(current);
-    irq_restore(flags);
-    thread_schedule(0);
-}
-
-struct task *pick_next_task()
-{
-    if (list_empty(&running_queue))
-    {
-        while (1)
-        {
-            uart_send_string("scheduler: run queue is empty\n");
-        };
-    }
-    struct task *next_task = list_first_entry(&running_queue, struct task, list);
-    unlink(&next_task->list);
-    insert_tail(&running_queue, &next_task->list);
-
-    return next_task;
-}
-////////////////////////////////////////////////////////////////////////////////////////////////
-//                                       thread                                               //
-////////////////////////////////////////////////////////////////////////////////////////////////
-
-static void switch_task(struct task *next)
-{
-    switch_to(&current->cpu_context, &next->cpu_context);
 }
 
 void thread_schedule(size_t _)
@@ -137,12 +36,8 @@ void thread_schedule(size_t _)
 
 void thread_init()
 {
-    struct task *cur = kmalloc(sizeof(struct task));
-    cur->state = TASK_RUNNING;
-    cur->pid = task_count++;
-    cur->need_resched = 0;
-    cur->timeout = get_current_time() + DEFAULT_TIMEOUT;
-    set_thread_ds(cur); // init the thread structure
+    struct task *init = create_task();
+    set_thread_ds(init); // init the thread structure
 }
 
 static void foo()
@@ -151,7 +46,7 @@ static void foo()
     {
         uart_printf("Thread id: %d i=%d\n", current->pid, i);
         // sleep_task(500);
-        delay(1000000);
+        delay(10000000);
     }
     kill_task(current, 0);
     thread_schedule(0);
@@ -159,17 +54,15 @@ static void foo()
 
 struct task *thread_create(void *func)
 {
-    struct task *new_task = kmalloc(sizeof(struct task));
-    new_task->kstack = kmalloc(0x2000);
-    new_task->state = TASK_RUNNING;
-    new_task->pid = task_count++;
-    new_task->need_resched = 0;
-    new_task->timeout = get_current_time() + DEFAULT_TIMEOUT;
-    new_task->cpu_context.lr = (unsigned long)func;
-    new_task->cpu_context.sp = (unsigned long)new_task->kstack + 0x2000 - sizeof(struct pt_regs);
+    struct task *new_thread = create_task();
+    new_thread->kstack = kmalloc(STACK_SIZE);
+    new_thread->state = TASK_RUNNING;
+    new_thread->cpu_context.lr = (unsigned long)func;
+    new_thread->cpu_context.sp = (unsigned long)new_thread->kstack + STACK_SIZE - sizeof(struct pt_regs);
 
-    add_task(new_task);
-    return new_task;
+    add_task(new_thread);
+
+    return new_thread;
 }
 
 void test_thread()
