+++ lab5/src/utils_c.c
@@ -6,7 +6,7 @@
     string part
 */
 
-int utils_str_compare(const char *a,const char *b)
+int utils_str_compare(const char *a, const char *b)
 {
     char aa, bb;
     do
@@ -21,10 +21,12 @@ int utils_str_compare(const char *a,const char *b)
     return aa - bb;
 }
 
-int utils_strncmp(const char *a, const char *b, size_t n) {
-  size_t i = 0;
-  while (i < n-1 && a[i] == b[i] && a[i] != '\0' && b[i] != '\0') i++;
-  return a[i] - b[i];
+int utils_strncmp(const char *a, const char *b, size_t n)
+{
+    size_t i = 0;
+    while (i < n - 1 && a[i] == b[i] && a[i] != '\0' && b[i] != '\0')
+        i++;
+    return a[i] - b[i];
 }
 void utils_newline2end(char *str)
 {
@@ -98,10 +100,12 @@ unsigned int utils_str2uint_dec(const char *str)
     return value;
 }
 
-size_t utils_strlen(const char *s) {
-  size_t i = 0;
-  while (s[i]) i++;
-  return i+1;
+size_t utils_strlen(const char *s)
+{
+    size_t i = 0;
+    while (s[i])
+        i++;
+    return i + 1;
 }
 
 /*
@@ -133,14 +137,15 @@ void cancel_reset()
 void align(void *size, size_t s)
 {
     unsigned int *x = (unsigned int *)size;
-    if ((*x) & (s-1))
+    if ((*x) & (s - 1))
     {
-        (*x) += s - ((*x) & (s-1));
+        (*x) += s - ((*x) & (s - 1));
     }
 }
 
-uint32_t align_up(uint32_t size, int alignment) {
-  return (size + alignment - 1) & -alignment;
+uint32_t align_up(uint32_t size, int alignment)
+{
+    return (size + alignment - 1) & -alignment;
 }
 
 void delay(unsigned int clock)
@@ -151,11 +156,13 @@ void delay(unsigned int clock)
     }
 }
 
-void memcpy(void *dst, const void *src, size_t n) {
-  char *_dst = dst;
-  const char *_src = src;
+void memcpy(void *dst, const void *src, size_t n)
+{
+    char *_dst = dst;
+    const char *_src = src;
 
-  while(n--) {
-    *_dst++ = *_src++;
-  }
+    while (n--)
+    {
+        *_dst++ = *_src++;
+    }
 }
